new file

+++ el&jstl_test/src/com/hyb/service/impl/UserServiceImpl.java
@@ -2,6 +2,7 @@ package com.hyb.service.impl;
 
 import com.hyb.dao.UserDao;
 import com.hyb.dao.impl.UserDaoImpl;
+import com.hyb.domain.PageBean;
 import com.hyb.domain.User;
 import com.hyb.service.UserService;
 
@@ -43,12 +44,49 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delSelectedUser(String[] ids) {
-        //遍历数组
-        for (String id : ids) {
-            //调用dao删除
-             dao.delete(Integer.parseInt(id));
+        if (ids != null && ids.length > 0) {
+            //遍历数组
+            for (String id : ids) {
+                //调用dao删除
+                dao.delete(Integer.parseInt(id));
+            }
         }
 
+
+    }
+
+    @Override
+    public PageBean<User> findUserByPage(String _currentPage, String _rows) {
+
+        int currentPage = Integer.parseInt(_currentPage);
+        int rows = Integer.parseInt(_rows);
+
+        if (currentPage <= 0) {
+            currentPage = 1;
+        }
+
+        //1.创建新的PageBean对象
+        PageBean<User> pb = new PageBean<User>();
+
+        //2.设置参数
+        pb.setCurrentPage(currentPage);
+        pb.setRows(rows);
+
+        //3.调用dao查询总记录数
+        int totalCount = dao.findTotalCount();
+        pb.setTotalCount(totalCount);
+
+        //4.调用dao查询list集合
+        //计算开始的索引
+        int start = (currentPage - 1) * rows;
+        List<User> list = dao.findByPage(start,rows);
+        pb.setList(list);
+
+        //5.计算总页码
+        int totalPage = totalCount % rows == 0 ? totalCount / rows : totalCount / rows + 1;
+        pb.setPageCount(totalPage);
+
+        return pb;
     }
